+++ xpasses/checks.go
@@ -2,6 +2,7 @@ package xpasses
 
 import (
 	"github.com/bflad/tfproviderlint/xpasses/XR001"
+	"github.com/bflad/tfproviderlint/xpasses/XR002"
 	"github.com/bflad/tfproviderlint/xpasses/XS001"
 	"golang.org/x/tools/go/analysis"
 )
@@ -11,5 +12,6 @@ import (
 // combining these Analyzers with additional custom Analyzers
 var AllChecks = []*analysis.Analyzer{
 	XR001.Analyzer,
+	XR002.Analyzer,
 	XS001.Analyzer,
 }